refactor: test setup
