+++ docs/e2e_test_setup.md
@@ -86,7 +86,7 @@ To run the demo, the following software needs to be installed.
    Example:
 
    ```bash
-   export E2E_CONFIG_PATH=/home/User/IBC-integration/e2e-config.yaml
+   export TEST_CONFIG_PATH=/home/User/IBC-integration/sample-config-archway.yaml
    export GOLOOP_IMAGE_ENV=goloop
    export GOLOOP_IMAGE_TAG_ENV=latest
    ```
