--- test/integration/README.md
@@ -1,5 +1,0 @@
-### Configuration
-
-```json
-
-```
