+++ test/testsuite/integrationsuite.go
@@ -3,91 +3,19 @@ package testsuite
 import (
 	"context"
 	"fmt"
-	"time"
-
 	conntypes "github.com/cosmos/ibc-go/v7/modules/core/03-connection/types"
 	chantypes "github.com/cosmos/ibc-go/v7/modules/core/04-channel/types"
-
-	interchaintest "github.com/icon-project/ibc-integration/test"
 	"github.com/icon-project/ibc-integration/test/chains"
-	"github.com/icon-project/ibc-integration/test/e2e/relayer"
-	"github.com/icon-project/ibc-integration/test/e2e/testconfig"
 	"github.com/strangelove-ventures/interchaintest/v7/ibc"
 	test "github.com/strangelove-ventures/interchaintest/v7/testutil"
 	"golang.org/x/sync/errgroup"
+	"time"
 )
 
 var (
 	rlyArgs     = []string{"--log-format", "json", "--debug", "--json"}
 	channelName = "ICON-ARCHWAY"
 )
-
-func (s *E2ETestSuite) SetupRelayer(ctx context.Context) (ibc.Relayer, error) {
-	config := testconfig.New()
-	chainA, chainB := s.GetChains()
-	r := relayer.New(s.T(), config.RelayerConfig, s.logger, s.DockerClient, s.network)
-	pathName := s.generatePathName()
-	ic := interchaintest.NewInterchain().
-		AddChain(chainA.(ibc.Chain)).
-		AddChain(chainB.(ibc.Chain)).
-		AddRelayer(r, "r").
-		AddLink(interchaintest.InterchainLink{
-			Chain1:  chainA.(ibc.Chain),
-			Chain2:  chainB.(ibc.Chain),
-			Relayer: r,
-			Path:    pathName,
-		})
-
-	eRep := s.GetRelayerExecReporter()
-	buildOptions := interchaintest.InterchainBuildOptions{
-		TestName:          s.T().Name(),
-		Client:            s.DockerClient,
-		NetworkID:         s.network,
-		BlockDatabaseFile: interchaintest.DefaultBlockDatabaseFilepath(),
-		SkipPathCreation:  true,
-	}
-	if err := ic.BuildChains(ctx, eRep, buildOptions); err != nil {
-		return nil, err
-	}
-	if err := chainA.BuildWallets(ctx, Owner); err != nil {
-		return nil, err
-	}
-	if err := chainB.BuildWallets(ctx, Owner); err != nil {
-		return nil, err
-	}
-	if err := chainA.BuildWallets(ctx, User); err != nil {
-		return nil, err
-	}
-	if err := chainB.BuildWallets(ctx, User); err != nil {
-		return nil, err
-	}
-	var err error
-	ctx, err = chainA.SetupIBC(ctx, Owner)
-	if err != nil {
-		return nil, err
-	}
-	ctx, err = chainB.SetupIBC(ctx, Owner)
-	if err != nil {
-		return nil, err
-	}
-	if err := ic.BuildRelayer(ctx, eRep, buildOptions); err != nil {
-		return nil, err
-	}
-	s.startRelayerFn = func(relayer ibc.Relayer) {
-		err := relayer.StartRelayer(ctx, eRep, pathName)
-		s.Require().NoError(err, fmt.Sprintf("failed to start relayer: %s", err))
-		s.T().Cleanup(func() {
-			if !s.T().Failed() {
-				if err := relayer.StopRelayer(ctx, eRep); err != nil {
-					s.T().Logf("error stopping relayer: %v", err)
-				}
-			}
-		})
-		s.Require().NoError(test.WaitForBlocks(ctx, 10, chainA.(ibc.Chain), chainB.(ibc.Chain)), "failed to wait for blocks")
-	}
-	s.relayer = r
-	return r, r.GeneratePath(ctx, eRep, chainA.(ibc.Chain).Config().ChainID, chainB.(ibc.Chain).Config().ChainID, pathName)
-}
 
 func (s *E2ETestSuite) CreateClient(ctx context.Context) error {
 	eRep := s.GetRelayerExecReporter()
