+++ test/testsuite/sample-config-archway.yaml
@@ -30,7 +30,7 @@ chains:
       name: archway
       chain_id: archway-1
       image:
-        repository: archwaynetwork/archwayd
+        repository: archwayd
         version: latest
         uid_gid: ""
       bin: archwayd
